Flytt kode der det brukes
i stedet for å ha en utilklasse med utydelig ansvar
